Changed protocol to support a Hashtable & bugfix
Hashtable allows for the sequence number to be sent along with the
update for indexing at the replica nodes. This way, we can retrieve
specific updates when needed (like when updating remote nodes after our
sequence number is > 1 more than it is.).

Bugfix fixed a sequence error caused by long delays or heavy traffic.